FIX: Use correct GLint type (rather than long) in ApplePBTextureTarget to resolve potential compilation issues if GLint is typedef'd to an incompatible type.  See: http://www.cegui.org.uk/phpBB2/viewtopic.php?f=3&t=4553

+++ cegui/src/RendererModules/OpenGL/CEGUIOpenGLApplePBTextureTarget.cpp
@@ -69,7 +69,7 @@ OpenGLApplePBTextureTarget::OpenGLApplePBTextureTarget(OpenGLRenderer& owner) :
                                 "CGLGetVirtualScreen failed: " +
                                 String(CGLErrorString(err)));
 
-    long fmt_count;
+    GLint fmt_count;
     CGLPixelFormatObj pix_fmt;
     if (err = CGLChoosePixelFormat(fmtAttrs, &pix_fmt, &fmt_count))
         throw RendererException("OpenGLApplePBTextureTarget - "
